refactor(ranking.beaten-games): migrate to Folio-native routing and render hook
